# app/api/monetization/route.ts
import { NextResponse } from 'next/server';

export async function POST(req: Request) {
    try {
        const authHeader = req.headers.get('authorization');
        if (authHeader !== `Bearer ${process.env.GPT_BRIDGE_TOKEN}`) {
            return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
        }

        // This is a placeholder for monetization analysis logic
        return NextResponse.json({
            success: true,
            opportunities: [
                { type: 'AdSense', status: 'Ready', suggestion: 'Add more vertical ad units to blog sidebar' },
                { type: 'Product', status: 'Active', suggestion: 'Create a bundle for AI voice interaction' },
                { type: 'Affiliate', status: 'Enabled', suggestion: 'Link ReadyPlayerMe avatars to partner deals' }
            ],
            currentStrategy: 'Hybrid: Ads + Direct Sales',
            timestamp: new Date().toISOString()
        });
    } catch (error: unknown) {
        return NextResponse.json({ error: 'Failed to analyze monetization' }, { status: 500 });
    }
}