# app/api/rollback/route.ts
import { NextResponse } from 'next/server';

export async function POST(req: Request) {
    try {
        const authHeader = req.headers.get('authorization');
        if (authHeader !== `Bearer ${process.env.GPT_BRIDGE_TOKEN}`) {
            return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
        }

        // Execute git rollback
        console.warn('Executing emergency rollback via Git...');

        // In a real environment, this would run:
        // execSync('git revert HEAD --no-edit && git push origin main');

        return NextResponse.json({
            success: true,
            message: 'Emergency rollback signal received and processed.',
            action: 'Git Revert HEAD',
            timestamp: new Date().toISOString()
        });
    } catch (error: unknown) {
        console.error('Rollback failed:', error);
        return NextResponse.json({ error: 'Rollback failed' }, { status: 500 });
    }
}